Working on a Windows buildable version in the hope of better debugging.

+++ cRio_Daq_cpp/trunk/cRio_DAQ/src/mythread.h
@@ -31,7 +31,7 @@ typedef HANDLE THREADHANDLE;
 typedef DWORD THREADID;
 
 #define DECLARETHREAD(starterName, className, functionName) \
-unsigned long __stdcall starterName (void* threadData) { \
+DWORD __stdcall starterName (void* threadData) { \
 	className* classReference = (className*) threadData; \
 	classReference->functionName(); \
 	return 0; \
